refac: carousel리스트를 전체 client가 아닌 carousel 부분만 클라이언트 컴포넌트로 분리
